+++ attributes/default.rb
@@ -1,5 +1,5 @@
 
-default['thruk']['version'] = '2.00-1'
+default['thruk']['version'] = '2.00-2'
 
 default['thruk']['major'] = node['platform_version'].to_i
 
@@ -29,27 +29,15 @@ default['thruk']['log_dir'] = '/var/log/thruk'
 default['thruk']['conf_dir'] = '/etc/thruk'
 default['thruk']['htpasswd'] = '/etc/thruk/htpasswd'
 default['thruk']['use_ssl'] = false
+default['thruk']['server_name'] = node['fqdn']
 
 default['thruk']['cert_name'] = node['fqdn']
+default['thruk']['cert_cookbook'] = 'thruk'
+
+default['thruk']['icon_cookbook'] = 'thruk'
+
 default['thruk']['cgi']['admin_group'] = 'admins'
 default['thruk']['cgi']['read_groups'] = 'all'
 default['thruk']['cgi']['lock_authors_names'] = 1
-
-default['thruk']['logo_path_prefix'] = '/thruk/icons/'
-default['thruk']['start_page'] = '/thruk/cgi-bin/tac.cgi'
-default['thruk']['first_day_of_week'] = 1
-default['thruk']['use_strict_host_authorization'] = 0
-default['thruk']['enable_shinken_features'] = 0
-
-default['thruk']['cmd_defaults']['ahas'] = 0
-default['thruk']['cmd_defaults']['broadcast_notification'] = 0
-default['thruk']['cmd_defaults']['force_check'] = 0
-default['thruk']['cmd_defaults']['force_notification'] = 0
-default['thruk']['cmd_defaults']['send_notification'] = 1
-default['thruk']['cmd_defaults']['sticky_ack'] = 1
-default['thruk']['cmd_defaults']['persistent_comments'] = 1
-default['thruk']['cmd_defaults']['persistent_ack'] = 0
-default['thruk']['cmd_defaults']['ptc'] = 0
-default['thruk']['cmd_defaults']['use_expire'] = 0
 
 default['thruk']['plugins'] = %w(dashboard minemap mobile panorama reports2 statusmap)