+++ test/fixtures/cookbooks/thruk_test/metadata.rb
@@ -1,0 +1,9 @@
+name             'thruk_test'
+maintainer       'Martha Greenberg'
+maintainer_email 'marthag@mit.edu'
+license          'Apache 2.0'
+description      'Integration tests for  thruk'
+long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
+version          '0.1.0'
+
+depends 'thruk'